src/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
decorator==4.4.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
PyYAML==5.3
six==1.14.0
traitlets==4.3.3
wcwidth==0.1.8
